move sample learners
