+++ clients/GVGAI-JavaClient/src/TestLearningClient.java
@@ -25,7 +25,7 @@ public class TestLearningClient
         } else {
             serverDir = "../..";
         }
-        String agentName = "agents.random.Agent";         //Agent to play with
+        String agentName = "sampleRandom.Agent";         //Agent to play with
         boolean visuals = false;
         /** Get arguments */
         Map<String, List<String>> params = new HashMap<>();
